requirements.txt
bottle==0.12.25
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
conan==1.59.0
fasteners==0.18
filelock==3.12.2
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
node-semver==0.6.1
numba==0.57.1
numpy==1.24.3
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
patch-ng==1.17.4
pluginbase==1.0.1
PyAudio==0.2.13
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
soundfile==0.12.1
SpeechRecognition==3.10.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
